requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
gunicorn==21.2.0
jinja2==3.1.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2
resend==0.8.0
python-dotenv==1.0.0
apscheduler==3.10.4
